Split unit tests into separate modules
